core/suggestion-service/requirements.txt
# Core dependencies
networkx==3.1.0  # Used for DAG operations
pyparsing==3.1.1  # Required by pyiceberg

# API dependencies (for future REST API implementation)
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1

# Utility libraries
typing-extensions==4.8.0  # Enhanced typing support
python-dotenv==1.0.0     # For loading environment variables

# LLM API clients
openai==1.76.0            # OpenAI API client
anthropic==0.50.0         # Anthropic API client